feat: refactor deployment code to handle all errors in one place
